Refactor names servlets
